Why redirect whole shell? Just redirect the loop.
And this is why I always have to wonder why people go through all of that trouble.  The only times I've *ever* needed another file-descriptor is when a script was when I was called by a program that passed in something useful for me to talk to.  Communicating with subshells or sibling programs is when I've used it the most.

As you can see here (http://unix.stackexchange.com/questions/18899/when-would-you-use-an-additional-file-descriptor), there aren't many non-lame uses for this sort of thing.  You'll know if you'll need it (and even then you'll be wrong half the time).